Bug 787502 - pre: Factor ByteArraysEntity and ByteArraysContentProducer out of Server11RepositorySession.
