+++ src/main/java/org/mozilla/gecko/sync/repositories/Server11RepositorySession.java
@@ -4,9 +4,6 @@
 
 package org.mozilla.gecko.sync.repositories;
 
-import java.io.IOException;
-import java.io.OutputStream;
-import java.io.UnsupportedEncodingException;
 import java.net.URI;
 import java.net.URISyntaxException;
 import java.util.ArrayList;
@@ -25,8 +22,9 @@ import org.mozilla.gecko.sync.Server11PreviousPostFailedException;
 import org.mozilla.gecko.sync.Server11RecordPostFailedException;
 import org.mozilla.gecko.sync.UnexpectedJSONException;
 import org.mozilla.gecko.sync.crypto.KeyBundle;
+import org.mozilla.gecko.sync.net.ByteArraysEntity;
 import org.mozilla.gecko.sync.net.SyncStorageCollectionRequest;
-import org.mozilla.gecko.sync.net.SyncStorageRequest;
+import org.mozilla.gecko.sync.net.SyncStorageRecordRequest;
 import org.mozilla.gecko.sync.net.SyncStorageRequestDelegate;
 import org.mozilla.gecko.sync.net.SyncStorageResponse;
 import org.mozilla.gecko.sync.net.WBOCollectionRequestDelegate;
@@ -37,31 +35,11 @@ import org.mozilla.gecko.sync.repositories.delegates.RepositorySessionStoreDeleg
 import org.mozilla.gecko.sync.repositories.delegates.RepositorySessionWipeDelegate;
 import org.mozilla.gecko.sync.repositories.domain.Record;
 
-import ch.boye.httpclientandroidlib.entity.ContentProducer;
-import ch.boye.httpclientandroidlib.entity.EntityTemplate;
-
 public class Server11RepositorySession extends RepositorySession {
-  private static byte[] recordsStart;
-  private static byte[] recordSeparator;
-  private static byte[] recordsEnd;
-
-  static {
-    try {
-      recordsStart    = "[\n".getBytes("UTF-8");
-      recordSeparator = ",\n".getBytes("UTF-8");
-      recordsEnd      = "\n]\n".getBytes("UTF-8");
-    } catch (UnsupportedEncodingException e) {
-      // These won't fail.
-    }
-  }
-
   public static final String LOG_TAG = "Server11Session";
 
-  private static final int UPLOAD_BYTE_THRESHOLD = 1024 * 1024;    // 1MB.
-  private static final int UPLOAD_ITEM_THRESHOLD = 50;
-  private static final int PER_RECORD_OVERHEAD   = 2;              // Comma, newline.
-  // {}, newlines, but we get to skip one record overhead.
-  private static final int PER_BATCH_OVERHEAD    = 5 - PER_RECORD_OVERHEAD;
+  protected static final int UPLOAD_BYTE_THRESHOLD = 1024 * 1024 - 512;    // 1MB minus some characters for wrapping in a JSON array.
+  protected static final int UPLOAD_ITEM_THRESHOLD = 50;
 
   /**
    * Used to track outstanding requests, so that we can abort them as needed.
@@ -299,7 +277,7 @@ public class Server11RepositorySession extends RepositorySession {
    * Access should be synchronized on <code>recordsBufferMonitor</code>.
    */
   protected ArrayList<String> recordGuidsBuffer = new ArrayList<String>();
-  protected int byteCount = PER_BATCH_OVERHEAD;
+  protected int byteCount = 0;
 
   @Override
   public void store(Record record) throws NoStoreDelegateException {
@@ -328,7 +306,7 @@ public class Server11RepositorySession extends RepositorySession {
       }
       recordsBuffer.add(json);
       recordGuidsBuffer.add(record.guid);
-      byteCount += PER_RECORD_OVERHEAD + delta;
+      byteCount += delta;
     }
   }
 
@@ -343,7 +321,7 @@ public class Server11RepositorySession extends RepositorySession {
 
       recordsBuffer = new ArrayList<byte[]>();
       recordGuidsBuffer = new ArrayList<String>();
-      byteCount = PER_BATCH_OVERHEAD;
+      byteCount = 0;
     }
   }
 
@@ -499,46 +477,6 @@ public class Server11RepositorySession extends RepositorySession {
       return;
     }
 
-    public class ByteArraysContentProducer implements ContentProducer {
-
-      ArrayList<byte[]> outgoing;
-      public ByteArraysContentProducer(ArrayList<byte[]> arrays) {
-        outgoing = arrays;
-      }
-
-      @Override
-      public void writeTo(OutputStream outstream) throws IOException {
-        int count = outgoing.size();
-        outstream.write(recordsStart);
-        outstream.write(outgoing.get(0));
-        for (int i = 1; i < count; ++i) {
-          outstream.write(recordSeparator);
-          outstream.write(outgoing.get(i));
-        }
-        outstream.write(recordsEnd);
-      }
-    }
-
-    public class ByteArraysEntity extends EntityTemplate {
-      private long count;
-      public ByteArraysEntity(ArrayList<byte[]> arrays, long totalBytes) {
-        super(new ByteArraysContentProducer(arrays));
-        this.count = totalBytes;
-        this.setContentType("application/json");
-        // charset is set in BaseResource.
-      }
-
-      @Override
-      public long getContentLength() {
-        return count;
-      }
-
-      @Override
-      public boolean isRepeatable() {
-        return true;
-      }
-    }
-
     public ByteArraysEntity getBodyEntity() {
       ByteArraysEntity body = new ByteArraysEntity(outgoing, byteCount);
       return body;
@@ -562,7 +500,7 @@ public class Server11RepositorySession extends RepositorySession {
       }
 
       URI u = serverRepository.collectionURI();
-      SyncStorageRequest request = new SyncStorageRequest(u);
+      SyncStorageRecordRequest request = new SyncStorageRecordRequest(u);
 
       request.delegate = this;
